+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 //import 'Projects/dice.dart';
 //import 'Projects/ball.dart';
 //import 'Projects/quizApp.dart';
-import 'Projects/story_brain.dart';
+import 'Projects/story_App.dart';
 
 void main() {
   runApp(const MyApp());
@@ -17,85 +17,6 @@ class MyApp extends StatelessWidget {
     return const MaterialApp(
       debugShowCheckedModeBanner: false,
       home: StoryPage(),
-    );
-  }
-}
-
-//Create a new storyBrain object from the StoryBrain class.
-StoryBrain storyBrain = StoryBrain();
-
-//
-class StoryPage extends StatefulWidget {
-  const StoryPage({Key? key}) : super(key: key);
-
-  @override
-  State<StoryPage> createState() => _StoryPageState();
-}
-
-class _StoryPageState extends State<StoryPage> {
-  //ALL MY FUNCT...
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: const BoxDecoration(
-          image: DecorationImage(
-              image: AssetImage('images/background.png'), fit: BoxFit.cover),
-        ),
-        padding: const EdgeInsets.symmetric(vertical: 50.0, horizontal: 15.0),
-        constraints: const BoxConstraints.expand(),
-        child: SafeArea(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Expanded(
-                flex: 12,
-                child: Center(
-                  child: Text(
-                    storyBrain.getStory(),
-                    style: const TextStyle(fontSize: 25.0),
-                  ),
-                ),
-              ),
-              Expanded(
-                flex: 2,
-                child: TextButton(
-                  onPressed: () {
-                    setState(() {
-                      storyBrain.nextStory(1);
-                    });
-                  },
-                  child: Text(
-                    storyBrain.getChoice1(),
-                    style: const TextStyle(fontSize: 20.0, color: Colors.white),
-                  ),
-                ),
-              ),
-              const SizedBox(
-                height: 20.0,
-              ),
-              Expanded(
-                flex: 2,
-                child: Visibility(
-                  child: TextButton(
-                    onPressed: () {
-                      setState(() {
-                        storyBrain.nextStory(2);
-                      });
-                    },
-                    child: Text(
-                      storyBrain.getChoice2(),
-                      style:
-                          const TextStyle(fontSize: 20.0, color: Colors.white),
-                    ),
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
